add Authentication and Authorization

+++ tasks/views.py
@@ -1,31 +1,42 @@
-from django.shortcuts import render, get_object_or_404, redirect
+from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth import login
+from django.contrib.auth.decorators import login_required
+from .forms import RegisterForm, TaskForm
 from .models import Task
-from .forms import TaskForm
 
-# 1. Список задач
+
+@login_required
 def task_list(request):
     tasks = Task.objects.all()
     return render(request, "tasks/task_list.html", {"tasks": tasks})
 
-# 2. Детальный просмотр задачи
+@login_required
 def task_detail(request, task_id):
     task = get_object_or_404(Task, id=task_id)
     return render(request, "tasks/task_detail.html", {"task": task})
 
-# 3. Создание задачи
+
+@login_required
 def task_create(request):
     if request.method == "POST":
         form = TaskForm(request.POST)
         if form.is_valid():
-            form.save()
+            task = form.save(commit=False)
+            task.author = request.user  
+            task.save()
             return redirect("task_list")
     else:
         form = TaskForm()
     return render(request, "tasks/task_form.html", {"form": form})
 
-# 4. Редактирование задачи
+
+@login_required
 def task_edit(request, task_id):
     task = get_object_or_404(Task, id=task_id)
+    
+    if task.author != request.user:  
+        return redirect("task_list")
+
     if request.method == "POST":
         form = TaskForm(request.POST, instance=task)
         if form.is_valid():
@@ -33,12 +44,33 @@ def task_edit(request, task_id):
             return redirect("task_list")
     else:
         form = TaskForm(instance=task)
+
     return render(request, "tasks/task_form.html", {"form": form})
 
-# 5. Удаление задачи
+
+@login_required
 def task_delete(request, task_id):
     task = get_object_or_404(Task, id=task_id)
+
+    if task.author != request.user:  # 
+        return redirect("task_list")
+
     if request.method == "POST":
         task.delete()
         return redirect("task_list")
+
     return render(request, "tasks/task_confirm_delete.html", {"task": task})
+
+
+def register(request):
+    if request.method == "POST":
+        form = RegisterForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            login(request, user)  # Автоматический вход после регистрации
+            return redirect("task_list")
+    else:
+        form = RegisterForm()
+    return render(request, "tasks/register.html", {"form": form})
+
+
